added helper for registering a funnel
